Flytter mapping av arenaVedtak inn i domeneobjekt for å ha det samlet på lik måte som andre FaktaGrunnlag
